requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
dj-database-url==3.1.0
dj-email-url==1.0.6
Django==6.0.1
django-cache-url==3.4.6
django-crispy-forms==2.5
django-filter==25.2
environs==14.5.0
idna==3.11
marshmallow==4.2.1
pillow==12.1.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
stripe==14.2.0
typing_extensions==4.15.0
urllib3==2.6.3
